+++ app/scripts/fhir/sdc-export-common.js
@@ -116,10 +116,13 @@ function addCommonSDCExportFns(ns) {
 
     // http://hl7.org/fhir/StructureDefinition/questionnaire-minOccurs
     if (targetItem.required) {
-      targetItem.extension.push({
-        "url" : "http://hl7.org/fhir/StructureDefinition/questionnaire-minOccurs",
-        "valueInteger" : parseInt(item.questionCardinality.min)
-      });
+      var minOccurInt = parseInt(item.questionCardinality.min);
+      if(minOccurInt > 1) {
+        targetItem.extension.push({
+          "url" : "http://hl7.org/fhir/StructureDefinition/questionnaire-minOccurs",
+          "valueInteger" : minOccurInt
+        });
+      }
     }
 
     // question repeats