+++ CHANGELOG.md
@@ -4,7 +4,12 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [20.1.2] 2019-12-31
 ### Fixed
-- Export calculatedExpression, variable and other FHIR extensions.
+- Export calculatedExpression, variable and other FHIR extensions in lforms format.
+
+## [20.1.2] 2019-12-26
+### Fixed
+- Exporting to FHIR no longer include the minOccurs extension if the value is 1
+  and "required" is true.
 
 ## [20.1.1] 2019-12-18
 ### Fixed